pyproject.toml
[project]
name = "data-analyse"
version = "0.1.0"
description = "Ferramenta de análise de dados com PandasAI e múltiplos backends de IA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.22.0",
    "pandasai>=2.0.0",
    "pandas>=2.0.0",
    "pymysql>=1.0.3",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0",
    "openpyxl>=3.1.0",
    "openai>=1.0.0",
    "langchain>=0.0.300",
    "langchain-community>=0.0.10",
    "langchain-ollama>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"